fixed a few ecal_mon_tui warnings

+++ app/mon/mon_tui/src/model/log.hpp
@@ -121,7 +121,7 @@ class LogModel
         return;
       }
       auto new_size = data.size() + new_entries_count;
-      int overflow_size = new_size - capacity;
+      size_t overflow_size = new_size - capacity;
 
       if(overflow_size > 0)
       {
